Linter not working on front end, updating and inserting movie working

+++ Opcao-01/back-end/models/Movie.js
@@ -3,16 +3,12 @@ const Movie = (sequelize, DataTypes) => {
     title: DataTypes.STRING,
     subtitle: DataTypes.STRING,
     storyline: DataTypes.STRING,
-    rating: DataTypes.INTEGER,
+    rating: DataTypes.DECIMAL(10, 2),
     imagePath: DataTypes.STRING,
     genre: DataTypes.STRING,
     director: DataTypes.STRING,
     quantity: DataTypes.INTEGER,
   }, { timestamps: false });
-
-  // user.associate = (models) => {
-  //   user.hasMany(models.BlogPost, { as: 'blogPosts', foreignKey: 'userId' });
-  // };
 
   return movie;
 };
